chore: moved sortByAlphabet and sortByWordOccurence to utils/common
